chore (refs T22614): move user Components to client/js.
