+++ client/js/bundles/procedure/administrationNewMemberListMastertoeblist.js
@@ -10,7 +10,7 @@
 /**
  * This is the entrypoint for administration_new_member_list_mastertoeblist.html.twig
  */
-import DpMasterToebList from '@DemosPlanUserBundle/components/DpMasterToebList'
+import DpMasterToebList from '@DpJs/components/user/DpMasterToebList'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
 
 const components = {
