+++ client/js/bundles/user/setPassword.js
@@ -14,7 +14,7 @@
 import dpValidate from '@DpJs/lib/validation/dpValidate'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
 import RegisterFlyout from '@DpJs/components/core/RegisterFlyout'
-import SetPassword from '@DemosPlanUserBundle/components/portalUser/SetPassword'
+import SetPassword from '@DpJs/components/user/portalUser/SetPassword'
 
 const components = {
   RegisterFlyout,
