+++ client/js/components/user/DpUserListExtendedItem.vue
